backend/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
asyncpg==0.31.0
pydantic==2.13.3
python-dotenv==1.2.2
python-multipart==0.0.26
